Split and reorder files, added new-dark-mode
