+++ css/about.css
@@ -1,0 +1,47 @@
+/* ========== About ========== */
+.about {
+    padding-top: 1rem;
+}
+
+.about__img {
+    width: 200px;
+    border-radius: 0.5rem;
+    margin-bottom: var(--mb-2);
+    display: inline-flex;
+    justify-self: center;
+}
+
+.about__description {
+    text-align: center;
+    margin-bottom: var(--mb-2-5);
+    color: var(--text-color);
+}
+
+.about__info {
+    display: flex;
+    justify-content: space-evenly;
+    margin-bottom: var(--mb-2-5);
+}
+
+.about__info-title {
+    font-size: var(--h2-font-size);
+    font-weight: var(--font-semi-bold);
+    color: var(--title-color);
+}
+
+.about__info-name {
+    font-size: var(--smaller-font-size);
+    font-family: var(--roboto-font);
+    color: var(--text-color);
+}
+
+.about__info-name,
+.about__info-title {
+    display: block;
+    text-align: center;
+}
+
+.about__buttons {
+    display: flex;
+    justify-content: center;
+}
