+++ css/generals.css
@@ -1,0 +1,278 @@
+@import url("https://fonts.googleapis.com/css2?family=Russo+One&display=swap");
+
+/* ========== Variables CSS ========== */
+:root {
+    --header-height: 3em;
+
+    /* ========== Colors ========== */
+    /*
+        Decide your color theme:
+        * Red: 0
+        * Green: 120
+        * Blue: 225
+        * Yellow: 60
+        * Orange: 30
+        * Purple: 285
+        More colors:
+            * https://en.wikipedia.org/wiki/Hue
+            * https://www.w3schools.com/colors/colors_hsl.asp
+    */
+    --hue-color: 150;
+
+    /* HSL color Light Mode */
+    --first-color: hsl(var(--hue-color), 75%, 25%);
+    --first-color-second: hsl(var(--hue-color), 50%, 75%);
+    --first-color-alt: hsl(var(--hue-color), 50%, 50%);
+    --first-color-lighter: hsl(var(--hue-color), 90%, 80%);
+
+    --title-color: hsl(var(--hue-color), 10%, 15%);
+    --text-color: hsl(var(--hue-color), 10%, 40%);
+    --mainText: hsl(var(--hue-color), 10%, 20%);
+    --text-color-light: hsl(var(--hue-color), 10%, 70%);
+    --input-color: hsl(var(--hue-color), 70%, 90%);
+    --body-color: hsl(var(--hue-color), 70%, 98%);
+    --container-color: hsl(var(--hue-color), 0%, 100% /* White */);
+    --scroll-bar-color: hsl(var(--hue-color), 15%, 85%);
+    --scroll-thumb-color: hsl(var(--hue-color), 10%, 70%);
+    --gradient-color: linear-gradient(
+        90deg,
+        rgba(32, 223, 128, 1) 0%,
+        rgba(0, 89, 3, 1) 50%,
+        rgba(32, 223, 128, 1) 100%
+    );
+
+    /* ========== Font & Typography ========== */
+    --maven-font: "Maven Pro", sans-serif;
+    --openSans-font: "Open Sans", sans-serif;
+    --body-font: "Poppins", sans-serif;
+    --roboto-font: "Roboto", sans-serif;
+    --robotoMono-font: "Roboto Mono", monospace;
+    --bigFat-font: "Russo One", sans-serif;
+
+    /* 1rem = 16px, .5rem = 8px */
+    --big-font-size: 2rem;
+    --h1-font-size: 1.5rem;
+    --h2-font-size: 1.25rem;
+    --h3-font-size: 1.125rem;
+    --normal-font-size: 0.875rem;
+    --small-font-size: 0.813rem;
+    --smaller-font-size: 0.75rem;
+
+    /* Font Weight */
+    --font-medium: 500;
+    --font-semi-bold: 600;
+    --font-bold: 700;
+
+    /* Margin Size - Bottom */
+    /* .25rem = 4px, .5rem = 8px, .75rem = 12px, 1rem = 16px, */
+    --mb-0-25: 0.25rem;
+    --mb-0-5: 0.5rem;
+    --mb-0-75: 0.75rem;
+    --mb-1: 1rem;
+    --mb-1-25: 1.25rem;
+    --mb-1-5: 1.5rem;
+    --mb-1-75: 1.75rem;
+    --mb-2: 2rem;
+    --mb-2-25: 2.25rem;
+    --mb-2-5: 2.5rem;
+    --mb-2-75: 2.75rem;
+    --mb-3: 3rem;
+    --mb-3-25: 3.25rem;
+    --mb-3-5: 3.5rem;
+    --mb-3-75: 3.75rem;
+
+    /* Z-Index */
+    --z-tooltip: 10;
+    --z-fixed: 100;
+    --z-modal: 1000;
+    /*  ======== Background  ========= */
+    --background-body-color: hsl(var(--hue-color), 66%, 99%);
+    --border-color: none;
+    --border-style: none;
+    --footer-color: var(--first-color);
+}
+
+/* =========== Variable Color Dark Theme ============ */
+body.dark-theme {
+    /* HSL color Dark Mode */
+    --first-color: hsl(var(--hue-color), 75%, 25% /* Not changed */);
+    --first-color-second: hsl(var(--hue-color), 30%, 10%);
+    --first-color-alt: hsl(var(--hue-color), 50%, 50% /* Not changed */);
+    --first-color-lighter: hsl(var(--hue-color), 90%, 80% /* Not changed */);
+
+    --title-color: hsl(var(--hue-color), 10%, 95%);
+    --text-color: hsl(var(--hue-color), 10%, 75%);
+    --mainText: hsl(var(--hue-color), 10%, 75%);
+    --text-color-light: hsl(var(--hue-color), 10%, 70% /* Not changed */);
+    --input-color: hsl(var(--hue-color), 30%, 15%);
+    --body-color: hsl(var(--hue-color), 1%, 15%);
+    --container-color: hsl(var(--hue-color), 30%, 15%);
+    --scroll-bar-color: hsl(var(--hue-color), 10%, 50%);
+    --scroll-thumb-color: hsl(var(--hue-color), 10%, 25%);
+    --gradient-color: linear-gradient(
+        90deg,
+        rgba(0, 89, 3, 1) 0%,
+        rgba(32, 223, 128, 1) 50%,
+        rgba(0, 89, 3, 1) 100%
+    );
+    --background-body-color: rgba(255, 255, 255, 0.068);
+    --border-color: rgba(255, 255, 255, 0.425);
+    --border-style: solid;
+    --footer-color: hsl(var(--hue-color), 53%, 18%);
+}
+
+/* ========== Media Queries ================= */
+/* Font Size for larger devices */
+@media screen and (min-width: 992px) {
+    :root {
+        /* 1rem = 16px, .5rem = 8px */
+        --big-font-size: 3rem;
+        --h1-font-size: 2.5rem;
+        --h2-font-size: 1.5rem;
+        --h3-font-size: 1.25rem;
+        --normal-font-size: 1rem;
+        --small-font-size: 0.875rem;
+        --smaller-font-size: 0.813rem;
+    }
+}
+/* ==========  ========== */
+
+/* ========== Nav Buttons Dark/Light ========== */
+.nav__btns {
+    display: flex;
+    align-items: center;
+    padding-left: var(--mb-0-5);
+    margin: 0;
+    margin-top: -5px;
+    gap: var(--mb-0-75);
+}
+
+.change-theme {
+    font-size: 1.25rem;
+    color: var(--title-color);
+    margin-right: var(--mb-1);
+    cursor: pointer;
+}
+
+.change-theme:hover {
+    color: var(--first-color);
+}
+
+/* ========== Base ========== */
+* {
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
+}
+
+html {
+    scroll-behavior: smooth;
+}
+
+body {
+    margin: 0 0 var(--header-height);
+    font-family: var(--body-font);
+    font-size: var(--normal-font-size);
+    background-color: var(--body-color);
+}
+
+h1,
+h2,
+h3,
+h4 {
+    color: var(--title-color);
+    font-weight: var(--font-semi-bold);
+}
+
+ul {
+    list-style: none;
+}
+
+a {
+    text-decoration: none;
+}
+
+img {
+    max-width: 100%;
+    height: auto;
+}
+
+/* ========== Reusable CSS classes ========== */
+.section {
+    padding: 2rem 1rem 4rem;
+    margin: auto;
+    min-height: 80vh;
+}
+
+.section__title {
+    font-size: var(--big-font-size);
+    font-family: var(--robotoMono-font);
+    color: var(--title-color);
+}
+
+.section__subtitle {
+    display: block;
+    font-size: var(--small-font-size);
+    margin-bottom: var(--mb-3);
+    color: var(--text-color);
+}
+
+.section__title,
+.section__subtitle {
+    text-align: center;
+}
+
+/* ========== Button Styling ========== */
+.button {
+    display: inline-block;
+    background-color: var(--first-color);
+    color: #fff;
+    padding: 1rem;
+    border-radius: 0.5rem;
+    font-weight: var(--font-medium);
+}
+.button:hover {
+    background-color: var(--first-color-alt);
+    color: #fff;
+}
+.button__icon {
+    font-size: 1.25rem;
+    margin-left: var(--mb-0-5);
+    transition: 0.3s;
+}
+
+.button--flex {
+    display: inline-flex;
+    align-items: center;
+}
+
+.button--white {
+    background-color: #fff;
+    color: var(--first-color);
+}
+
+.button--white:hover {
+    color: var(--first-color-lighter);
+}
+
+/* ========== Layout ========== */
+.container {
+    padding: 0;
+    margin: 0;
+    max-width: 100%;
+}
+
+.grid {
+    display: grid;
+}
+
+.header {
+    width: 100%;
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    z-index: var(--z-fixed);
+    background-color: var(--body-color);
+    margin-right: var(--mb-0-25);
+}
