+++ css/home.css
@@ -1,0 +1,167 @@
+/* ========== Home ========== */
+
+.home {
+    height: 85vh;
+    align-content: space-between;
+    margin: 3rem 0 2rem 0;
+    padding: 6rem 0 2rem;
+}
+
+.home__container {
+    display: flex;
+    padding: 1rem;
+    justify-content: space-around;
+    align-items: center;
+}
+
+.home__content {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-areas: "home__social home__content";
+    padding-top: var(--mb-2-25);
+    align-items: center;
+    height: auto;
+    align-content: space-between;
+    margin-right: 0;
+}
+
+.home__social {
+    display: grid;
+    grid-area: home__social;
+    grid-template-columns: max-content;
+    row-gap: 0.5rem;
+    column-gap: 0;
+    margin: 0;
+}
+
+.home__social-icon {
+    font-size: 1.5rem;
+    color: var(--first-color);
+    border-radius: 100%;
+    padding: 0.5rem;
+    height: 48px;
+    width: 48px;
+    text-align: center;
+    background: var(--body-color);
+}
+
+.home__social-icon:hover {
+    color: var(--first-color-alt);
+    box-shadow: 0 0 0.5em rgb(0 0 0 / 50%);
+    background: var(--background-body-color);
+}
+
+.home__data {
+    grid-area: home__content;
+    align-content: center;
+    text-align: center;
+    display: block;
+    width: 100%;
+}
+
+.home__welcome {
+    text-align: center;
+    font-size: var(--small-font-size);
+    color: var(--text-color);
+    padding: 1.5rem 0;
+    display: block;
+}
+
+.home__title {
+    text-align: center;
+    font-size: var(--big-font-size);
+    color: var(--title-color);
+    margin: 1rem 0 1rem 0;
+}
+#typewriter {
+    min-height: 5rem;
+}
+.Typewriter__wrapper__title {
+    font-family: var(--maven-font);
+}
+.Typewriter__cursor__title {
+    font-size: var(--big-font-size);
+    color: var(--first-color-alt);
+    font-family: var(--maven-font);
+}
+
+.Typewriter__wrapper {
+    text-align: center;
+    font-size: var(--h2-font-size);
+    font-family: var(--robotoMono-font);
+    font-weight: var(--font-medium);
+    color: var(--text-color);
+}
+.Typewriter__cursor {
+    font-size: var(--h2-font-size);
+    color: var(--first-color-alt);
+    font-family: var(--maven-font);
+    font-weight: 600;
+}
+
+.home__subtitle {
+    text-align: center;
+    font-size: var(--h3-font-size);
+    color: var(--text-color);
+    font-weight: var(--font-medium);
+    margin: 1rem 0 1rem 0;
+    padding: 1rem 1.5rem 1rem 1.5rem;
+}
+
+/* Scroll Down */
+@keyframes scrollDown {
+    0% {
+        margin-top: 0px;
+        opacity: 0;
+    }
+    25% {
+        margin-top: 7.5px;
+        opacity: 1;
+    }
+    50% {
+        margin-top: 15px;
+        opacity: 1;
+    }
+    75% {
+        margin-top: 22.5px;
+        opacity: 1;
+    }
+    100% {
+        margin-top: 30px;
+        opacity: 0;
+    }
+}
+
+.home__scroll {
+    display: flex;
+    justify-content: center;
+    animation: scrollDown 3s infinite;
+    padding-top: 3rem;
+}
+
+.home__scroll-button {
+    display: grid;
+    text-align: center;
+    color: var(--first-color);
+    transition: 0.5s;
+}
+
+.home__scroll-button:hover {
+    color: var(--first-color-alt);
+    transform: translateY(0.25rem);
+}
+
+.home__scroll-mouse {
+    font-size: 2rem;
+}
+
+.home__scroll-name {
+    font-size: var(--small-font-size);
+    color: var(--title-color);
+    font-weight: var(--font-medium);
+    margin: var(--mb-0-25);
+}
+
+.home__scroll-arrow {
+    font-size: var(--h3-font-size);
+}
