+++ css/mediaQueries.css
@@ -1,0 +1,282 @@
+
+/* ========== Media Queries ========== */
+/* Small Screen */
+@media screen and (max-width: 350px) {
+    .nav__menu {
+        padding: 2rem 0.25rem 4rem;
+    }
+
+    .nav__logo,
+    .nav__toggle {
+        font-size: 1rem;
+    }
+
+    .nav__list {
+        column-gap: 0;
+    }
+    .home {
+        margin: 0;
+        padding: 0;
+    }
+    .home__container {
+        padding: 0.5rem;
+    }
+
+    .home__content {
+        grid-template-columns: none;
+    }
+    .home__welcome {
+        padding: 0.75rem 0 0;
+    }
+    .home__subtitle {
+        font-size: var(--h2-font-size);
+        margin: 0;
+        padding: 0 0.75rem 0.75rem 0;
+    }
+
+    .home__scroll {
+        margin-top: var(--mb-0-25);
+    }
+
+    .home__scroll-mouse {
+        margin-top: var(--mb-0-25);
+    }
+
+    .about {
+        margin-top: 6rem;
+    }
+    .qualifications {
+        padding: 2rem 1rem 0;
+    }
+    .qualification__data {
+        gap: 0.5rem;
+    }
+    .py-2 {
+        padding: 0 0.25rem 0.5rem !important;
+    }
+    .col-auto {
+        padding-right: 5px;
+        padding-left: 5px;
+    }
+    .contact_mail {
+        margin-left: var(--mb-1-5);
+    }
+
+    .footer__bg {
+        padding: 0.5rem;
+    }
+    .footer__copy {
+        margin: 0;
+    }
+
+    ::-webkit-scrollbar {
+        width: 0.4rem;
+        border-radius: 0.4rem;
+    }
+
+    ::-webkit-scrollbar-thumb {
+        background-color: var(--scroll-thumb-color);
+        border-radius: 0.2rem;
+    }
+}
+
+/* Medium size Screen devices */
+@media screen and (min-width: 568px) {
+    .home__content {
+        grid-template-columns: max-content 1fr;
+    }
+    .section {
+        min-height: 60vh;
+    }
+    .home__data {
+        grid-column: initial;
+    }
+
+    .about__container,
+    .skill__container,
+    .portfolio__container,
+    .contact__container,
+    .footer__container {
+        grid-template-columns: repeat(1, 1fr);
+        padding-left: var(--mb-1);
+        padding-right: var(--mb-1);
+    }
+
+    /* .qualification__sections{
+        display: grid;
+        grid-template-columns: 0.85fr;
+        justify-content: center;
+      } */
+
+    .portfolio {
+        padding-bottom: var(--mb-0-25);
+    }
+    .contact {
+        padding-top: 0;
+    }
+    .qualification__start-icon {
+        position: absolute;
+        transform: translateY(60%);
+    }
+}
+
+/* Large size Screen devices */
+@media screen and (min-width: 768px) {
+    .container {
+        margin-left: auto;
+        margin-right: auto;
+    }
+
+    body {
+        margin: 0;
+    }
+
+    .section {
+        padding: 6rem 0 2rem;
+    }
+
+    .section__subtitle {
+        margin-bottom: 4rem;
+    }
+
+    .header {
+        top: 0;
+        bottom: initial;
+    }
+
+    .header,
+    .main,
+    .footer__container {
+        padding: 0 1rem;
+    }
+
+    .nav {
+        height: calc(var(--header-height) + 1.5rem);
+        column-gap: 1rem;
+    }
+
+    .nav__icon,
+    .nav__close,
+    .nav__toggle {
+        display: none !important;
+    }
+
+    .nav__icon {
+        display: none !important;
+    }
+
+    .nav__list {
+        display: flex;
+        column-gap: 2rem;
+    }
+
+    .nav__menu {
+        margin-left: auto;
+    }
+
+    .change-theme {
+        margin: 0;
+        /* padding: 0; */
+    }
+
+    .home__container {
+        row-gap: 5rem;
+    }
+
+    .home__content {
+        padding-top: 2.5rem;
+        column-gap: 2rem;
+    }
+
+    .home__scroll-button {
+        margin-left: 3rem;
+    }
+
+    .about {
+        margin-top: 4rem;
+    }
+    .about__container {
+        column-gap: 5rem;
+    }
+
+    .about__img {
+        width: 350px;
+    }
+
+    /* .about__description {
+    text-align: initial;
+  } */
+
+    /* .about__info{
+      justify-content: space-between;
+    } */
+
+    .about__buttons {
+        justify-content: center;
+    }
+
+    .qualification__button {
+        margin: 0 var(--mb-1);
+    }
+
+    .footer__container {
+        grid-template-columns: repeat(3, 1fr);
+    }
+
+    .footer__bg {
+        padding: 3rem 0 3.5rem;
+    }
+
+    .footer__links {
+        flex-direction: row;
+        column-gap: 2rem;
+    }
+
+    .footer__socials {
+        justify-content: flex-end;
+    }
+
+    .footer__copy {
+        margin-top: 4.5rem;
+    }
+
+    /* .home__data{
+      transform: translateX(80px);
+    } */
+}
+
+/* Larger size Screen devices */
+@media screen and (min-width: 1024px) {
+    .header,
+    .main,
+    .footer__container {
+        padding: 0;
+    }
+    .nav__icon {
+        display: none !important;
+    }
+    #home {
+        height: 85vh;
+    }
+
+    .contact__form {
+        width: 460px;
+    }
+
+    .contact__inputs {
+        grid-template-columns: repeat(2, 1fr);
+    }
+
+    .contact__container {
+        padding-left: var(--mb-3);
+        padding-top: var(--mb-3);
+        row-gap: 4rem;
+    }
+
+    .form {
+        min-width: 700px;
+    }
+    .footer__container {
+        padding: 0 3rem;
+    }
+}
