+++ css/new-dark-mode.css
@@ -1,0 +1,68 @@
+/* The switch - the box around the slider */
+
+.switch {
+    display: block;
+    float: right;
+    margin: 3rem;
+    --width-of-switch: 3.5rem;
+    --height-of-switch: 2rem;
+    /* size of sliding icon -- sun and moon */
+    --size-of-icon: 1.4rem;
+    /* it is like a inline-padding of switch */
+    --slider-offset: 0.3rem;
+    position: relative;
+    width: var(--width-of-switch);
+    height: var(--height-of-switch);
+}
+
+/*  Hide default HTML checkbox */
+.switch input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+}
+
+/* The slider */
+.slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: #f4f4f5;
+    -webkit-transition: .4s;
+    transition: .4s;
+    border-radius: 30px;
+}
+
+.slider:before {
+    position: absolute;
+    content: "";
+    height: var(--size-of-icon, 1.4rem);
+    width: var(--size-of-icon, 1.4rem);
+    border-radius: 50%;
+    left: var(--slider-offset, 0.3rem);
+    top: 50%;
+    transform: translateY(-50%);
+    background: linear-gradient(40deg, #ff0080, #ff8c00 70%);
+    -webkit-transition: .4s;
+    transition: .4s;
+}
+
+input:checked + .slider {
+    background-color: #303136;
+}
+
+input:checked + .slider:before {
+    left: calc(100% - (var(--size-of-icon, 1.4rem) + var(--slider-offset, 0.3rem)));
+    background: #303136;
+    /* Change the value of second inset in box-shadow to change the angle and direction of the moon */
+    box-shadow: inset -3px -2px 5px 2px #8983f7, inset -10px -4px 0 0 #a3dafb;
+}
+
+/* Dark theme */
+.dark-mode-theme {
+    background: #212121;
+    color: #e6e6e6;
+}
