+++ css/qualifications.css
@@ -1,0 +1,164 @@
+
+/* ========== Qualifications ========== */
+.qualification__content {
+    max-width: 1400px;
+    min-height: 850px;
+    margin: 0 auto;
+}
+
+.qualification__button {
+    color: var(--first-color);
+    font-size: var(--h2-font-size);
+    font-weight: var(--font-medium);
+    cursor: pointer;
+    position: relative;
+    transform: translateX(15%);
+    margin: 1rem 0 0.5rem 0;
+}
+
+.qualification__button:hover {
+    color: var(--first-color-alt);
+}
+
+.qualification__icon {
+    font-size: var(--h1-font-size);
+    margin-right: 0.5rem;
+}
+
+.qualification__journey {
+    display: flex;
+    justify-content: space-between;
+}
+
+.qualification__line {
+    display: block;
+    width: 1.5px;
+    height: 100%;
+    background-color: var(--first-color-lighter);
+    align-items: flex-end;
+    transform: translateX(20px);
+    margin-right: 0;
+    padding: 0;
+}
+
+.qualification__calendar {
+    font-size: var(--small-font-size);
+    color: var(--text-color-light);
+}
+.qualification__calendar-icon {
+    font-size: var(--normal-font-size);
+}
+
+.qualification__title {
+    font-size: var(--h3-font-size);
+    font-weight: var(--font-medium);
+}
+
+.qualification__subtitle {
+    display: inline-block;
+    font-size: var(--small-font-size);
+    margin: 0 0 0.5rem;
+    color: var(--text-color);
+}
+
+.transparent_card {
+    background-color: var(--background-body-color);
+    border-color: var(--border-color);
+}
+
+.mirror-face {
+    border: 1px solid #fff;
+    border-radius: 15px;
+    font-size: 2rem;
+    height: initial;
+    overflow: hidden;
+    position: relative;
+    padding: 0;
+    margin: 0;
+    border-color: var(--first-color);
+}
+
+.task {
+    position: relative;
+    overflow: hidden;
+    cursor: pointer;
+    perspective: 800px;
+    transform-style: preserve-3d;
+}
+
+.abstract {
+    transition: 0.8s ease all;
+}
+
+.details {
+    /* display: flex;
+  justify-content: space-around; */
+    max-height: 0;
+    padding: 1rem 0 0;
+    overflow: hidden;
+    visibility: hidden;
+    transform: rotateX(-180deg);
+    transform-origin: top center;
+    -webkit-backface-visibility: hidden;
+    backface-visibility: hidden;
+    transition: 0.8s transform ease;
+    display: grid;
+}
+
+.details:before {
+    content: "";
+    display: block;
+    position: absolute;
+    top: 0;
+    left: 10%;
+    right: 10%;
+    height: 1px;
+    background: var(--text-color-light);
+}
+.certificate__button {
+    width: 100%;
+    text-align: center;
+    font-size: var(--small-font-size);
+    padding: 10px;
+    border-radius: 10px;
+    cursor: pointer;
+    display: inline-block;
+    margin: 0 auto;
+    transition: all 0.5s;
+}
+.certificate__button:hover {
+    box-shadow: 0 2.8px 2.2px rgba(0, 0, 0, 0.068),
+        0 6.7px 5.3px rgba(0, 0, 0, 0.092);
+}
+
+.certificate__button:hover i {
+    padding: 0 0 0 15px;
+    transition: all 0.5s;
+}
+.certificate__icon {
+    padding: 0 0.5rem;
+}
+
+.task:hover .details {
+    max-height: none;
+    overflow: hidden;
+    visibility: visible;
+    transform: rotateX(0deg);
+    transition: 0.8s transform ease;
+}
+
+.qualification__start {
+    display: inline-flex;
+    width: 14px;
+    height: 14px;
+    margin-top: var(--mb-0-5);
+    justify-content: flex-end;
+    background-color: transparent;
+    border-radius: auto;
+    transform: translate(15px, 10px);
+}
+
+.qualification__start-icon {
+    font-size: var(--h3-font-size);
+    color: var(--mainText);
+}
