+++ css/scrollUp.css
@@ -1,0 +1,27 @@
+/* ========== Scroll up ========== */
+
+.scroll_up {
+    position: fixed;
+    right: 1rem;
+    bottom: -20%;
+    background-color: var(--first-color);
+    opacity: 0.8;
+    border-radius: 0.4rem;
+    z-index: var(--z-tooltip);
+    transition: 0.4s;
+}
+
+.scroll_up:hover {
+    background-color: var(--first-color-alt);
+}
+
+.scroll_up-icon {
+    font-size: 1.25rem;
+    color: var(--body-color);
+    padding: 0.5rem;
+}
+
+/* Show Scroll */
+.show-scroll {
+    bottom: 5rem;
+}
