+++ js/sectionHighlight.js
@@ -1,0 +1,26 @@
+/* ========== Scroll Section active highlight ========== */
+
+const sections = document.querySelectorAll("section[id]");
+
+function scrollActive() {
+    const scrollY = window.pageYOffset;
+
+    sections.forEach((current) => {
+        const sectionHeight = current.offsetHeight;
+        const sectionTop = current.offsetTop - 50;
+        sectionId = current.getAttribute("id");
+
+        if (scrollY > sectionTop && scrollY <= sectionTop + sectionHeight) {
+            document
+                .querySelector(".nav__menu a[href*=" + sectionId + "]")
+                .classList.add("active-link");
+        } else {
+            document
+                .querySelector(".nav__menu a[href*=" + sectionId + "]")
+                .classList.remove("active-link");
+        }
+    });
+
+    console.log("Section highlight is working!");
+}
+window.addEventListener("scroll", scrollActive);
