+++ new-dark-mode.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" href="./css/new-dark-mode.css">
+</head>
+<body>
+    <label class="switch">
+        <input type="checkbox" id="toggle-btn">\
+        <span class="slider"></span>
+    </label>
+    <script src="/js/new-dark-mode.js"></script>
+</body>
+</html>
