+++ script.js
@@ -1,0 +1,5 @@
+import "./js/darkMode.js";
+import "./js/menu.js";
+import "./js/typewriter.js";
+import "./js/swiper.js";
+import "./js/scrollUp.js";
