+++ style.css
@@ -1,0 +1,12 @@
+@import "/css/generals.css";
+@import "/css/header.css";
+@import "/css/home.css";
+@import "/css/about.css";
+@import "/css/skills.css";
+@import "/css/qualifications.css";
+@import "/css/projects.css";
+@import "/css/contact.css";
+@import "/css/footer.css";
+@import "/css/mediaQueries.css";
+@import "/css/scrollUp.css";
+@import "/css/scrollBar.css";
